Enable gosec for security linting (#294)
* Enable gosec for security linting

gosec does static analysis on the go code-base. it's already a linter in
golangci-lint, so enabling it is quite simple.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>

* Appease gosec warnings

This appeases the two prominent warnings:

* Potential hardcoded secret
* Implicit memory aliasing in for loop

While these were actually false positives, it's important to address
these so we don't introduce actual issues in the future.

For the hardcoded secrets warning, we simply added `nolint`. The
warnings were not actual hardcoded secrets, but instead configuration
settings.

For the "implicit memory aliasing" warning, we stopped using the value
of the for-loop copy, and instead used the original reference. This way,
we don't use copies and even slightly improve performance.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>