Migrating RBAC Scanner to Kubernetes
With defsec migrating rbac scanner to kubernetes scanner, Trivy-Operator should only
use the kubernetes scanner to handler all kubernetes resources.
Since, defsec has not released a tag with these changes yet, we are updating the
defsec dependency to point to the commit we need. We can update this once defsec
releases a new tag.
We have fixed a unit test case that was depending on rbac esources.