fix!: Helm chart: add missing quote filter in config template (#1015)
The values for trivy.resources were not quoted.
If the Helm value is a string containing only a number,
e.g. `trivy.resources.requests.cpu: "2"`,
the generated YAML for the config map is
`trivy.resources.requests.cpu: 2`, which is of
the type number, not string, which is not a valid
value for a Kubernetes config map.