preserve unknown fields in vulnerabilityreport (#1455)
I built a custom operator (https://github.com/telekom-mms/trivy-dojo-report-operator)
that uses the kopf-operator (github.com/nolar/kopf/) framework.
This framework can act on kubernetes-resource creation. To detect which resources
were already worked on, kopf adds a the last state of the resource as an
annotation or status to the resource.I cannot use the annotation-feature, because
of this issue: https://github.com/kubernetes-sigs/kubebuilder/issues/2556 (annotations
get too long).

So my proposal is to preseve unknown fields in the vulnerabilityreport-crd. This way
I can add a status-field which kopf then can use to store the state.