Fix: Added possible fix for missing configFile configuration. Not testes yet.
Added volume and volume mount checks, which should then append the found volumes to their correct path - a feature which has not been working yet.

configFile argument seemed to never create and populate the path /etc/trivy/trivy-config.yaml. There was missing configuration in the helm chart (trivy.yaml) and supposedly in the allocation of an exisiting configFile (in the image and filesystem plugin for trivy)